requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
google-generativeai
duckduckgo-search
slowapi
python-dotenv
httpx
jinja2
python-multipart
